+++ project-chashi/src/main/java/com/io/yy/marketing/vo/CsMembercardConsumQueryVo.java
@@ -44,7 +44,7 @@ public class CsMembercardConsumQueryVo implements Serializable {
     private Integer consumType;
 
     @ApiModelProperty(value = "减免时长")
-    private Integer consumTime;
+    private Double consumTime;
 
     @ApiModelProperty(value = "减免金额")
     private Double consumAmount;
